fix(geometry): recover dense faceted-reveal void cuts (Sloan traversal + volume-safe batch)
The residual tearing on the densest ISSUE_098 Poroton walls (wall
3FceP9AqX1_92g5eDdrV5C: 7 faceted-BREP splayed windows on one face) came from
two kernel limits the off-plane fix (#1658) did not reach:

1. The incremental channel recovery bails on a self-touching 'figure-8' channel
   (one boundary vertex visited twice) that its single-successor boundary walk
   cannot traverse, leaving constraints unforced.
2. Any unrecovered constraint hard-rejected the whole batched N-ary cut to the
   sequential per-cutter path, which re-jitters and re-fragments the wall.

Fix 1 - recover_via_traversal: a robust Sloan ordered-segment traversal that
walks the crossed triangles from a to b (ordered ALONG the segment, so immune to
the pinch) and rebuilds the two chains, forcing the edge. Run as a fallback when
the fast boundary-walk recovery bails, so the common path stays byte-identical.

Fix 2 - difference_all_lenient + volume-safe subtract_many: keep the exact
(cleaner) batched cut when its removed volume matches a sequential reference;
only fall back to sequential when the non-conforming batch would over/under-cut
(guards the #559171 / #1167 volume family).

Result: wall 3FceP9AqX1... ~2072 -> ~108 open edges, removed volume unchanged;
wall W stays 42; #1167 rotated-wall under-cut and the full CSG corpus stay green.

Refactor to hold the module-size ratchet: constraint recovery (phase D) split
into kernel/retriangulate_recover.rs; signed_volume6 into kernel/signed_volume.rs.
