+++ rust/geometry/src/kernel/arrangement/mod.rs
@@ -36,7 +36,7 @@ mod tests;
 
 pub use self::boolean::{
     boolean, boolean_manifest, boolean_topology_hash, box_mesh, cube_mesh, difference_all,
-    union_all,
+    difference_all_lenient, union_all,
 };
 
 pub type Tri = [[f64; 3]; 3];
