+++ rust/geometry/src/kernel/mod.rs
@@ -25,8 +25,10 @@ pub mod manifest;
 pub mod mesh_bridge;
 pub mod predicates;
 pub mod rational;
+mod signed_volume;
 pub mod retriangulate;
 mod retriangulate_cleanup;
+mod retriangulate_recover;
 pub mod tritri;
 
 /// Three-valued exact sign.
